EEPROM: simulate a ~6 ms device busy period after writing a single unit

+++ src/nba/include/nba/save_state.hpp
@@ -258,7 +258,7 @@ struct SaveState {
     } flash;
 
     struct EEPROM {
-      u8 state;
+      u16 state;
       u16 address;
       u64 serial_buffer;
       u8 transmitted_bits;
